Solution to Input Two Integers and Print Their Sum

+++ Input Two Integers and Print Their Sum@@6788d5012db25efec946f93d/code.c
@@ -3,7 +3,7 @@ int main()
 {
     int a,b;
     int c;
-    c=(a+b);
+    c=(int)(a+b);
     scanf("%d%d",&a,&b);
     printf("sum: %d",c);
 }
